build(): When targets is a unary list, return a unary tuple
Hi there,

I’d like to propose this subtle but breaking change in behavior. In our library which consumes artifax, we have an optional `targets` parameter, which is either a list or None.

We pass this list into Artifax, and it would be helpful if Artifax behaved the same regardless of the length of that list.

The current behavior is that `build(targets=[“a”, “b”])` returns a tuple, while `build(targets=[“a”])` and `build(targets=“a”)` return a bare result.

With this change, a list of targets always returns a tuple of results. When the caller wants a bare result, pass `targets` as a string.

Thanks for considering! If you’d like to see how we’re using this, I’m happy to share in more detail.